Tests/LibWeb/Text/expected/HTML/window-or-worker-global-origin.txt

window.origin is a string: true
window.origin matches location.origin: true
worker.origin is a string: true
worker.origin matches location.origin: true
